Fix compile error in single-file compilation mode.
I have to compile in single-file mode due to an obscure compiler crash on macOS. For some reason, the global symbols within the `static if` are not visible when the matching static if within `GLLoader` is compiled. Since they are `private` anyway, just removing the outer `static if` seems like a viable workaround.